cpe3500/lab2/python_plots/pyproject.toml
[project]
name = "python-plots"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arm-gdb>=0.9.10",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
]
